backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pillow
aiofiles
python-multipart
